cloud_functions/home/requirements.txt
# Function dependencies, for example:
# package>=version
pytz
jinja2
google-cloud-datastore
